refactor(archiver): extract l1 synchronizer
