+++ yarn-project/txe/src/state_machine/archiver.ts
@@ -4,17 +4,9 @@ import { CheckpointNumber, type EpochNumber, type SlotNumber } from '@aztec/foun
 import { Fr } from '@aztec/foundation/curves/bn254';
 import type { EthAddress } from '@aztec/foundation/eth-address';
 import type { AztecAsyncKVStore } from '@aztec/kv-store';
-import {
-  type CheckpointId,
-  L2Block,
-  type L2BlockId,
-  type L2TipId,
-  type L2Tips,
-  type ValidateCheckpointResult,
-} from '@aztec/stdlib/block';
-import { Checkpoint, type PublishedCheckpoint } from '@aztec/stdlib/checkpoint';
+import type { CheckpointId, L2BlockId, L2TipId, L2Tips, ValidateCheckpointResult } from '@aztec/stdlib/block';
+import type { PublishedCheckpoint } from '@aztec/stdlib/checkpoint';
 import type { L1RollupConstants } from '@aztec/stdlib/epoch-helpers';
-import type { BlockHeader } from '@aztec/stdlib/tx';
 
 /**
  * TXE Archiver implementation.
@@ -92,18 +84,6 @@ export class TXEArchiver extends ArchiveSourceBase {
 
   public getL2EpochNumber(): Promise<EpochNumber | undefined> {
     throw new Error('TXE Archiver does not implement "getL2EpochNumber"');
-  }
-
-  public getCheckpointsForEpoch(_epochNumber: EpochNumber): Promise<Checkpoint[]> {
-    throw new Error('TXE Archiver does not implement "getCheckpointsForEpoch"');
-  }
-
-  public getBlocksForEpoch(_epochNumber: EpochNumber): Promise<L2Block[]> {
-    throw new Error('TXE Archiver does not implement "getBlocksForEpoch"');
-  }
-
-  public getBlockHeadersForEpoch(_epochNumber: EpochNumber): Promise<BlockHeader[]> {
-    throw new Error('TXE Archiver does not implement "getBlockHeadersForEpoch"');
   }
 
   public isEpochComplete(_epochNumber: EpochNumber): Promise<boolean> {
